refactor(admin): remove redundant CatalogUiPrimitives pass-through file
- Remove empty pass-through file CatalogUiPrimitives.tsx
- Update catalog component import paths to import directly from primitives
- Verify type check and production build pass with zero errors
