--- apps/admin/src/components/catalog/CatalogUiPrimitives.tsx
@@ -1,2 +1,0 @@
-"use client";
-export * from "./primitives";
